requirements.txt
streamlit
google-generativeai
python-docx
markdown
htmldocx